Greatly improve UI with Material 3, Cards, and Lists. Refactor code.
- Refactored `main.dart` into `home_page.dart`, `settings_page.dart`, and `clickable_link.dart`.
- Updated `main.dart` to use Material 3 and a seed color.
- Overhauled `home_page.dart` to use a Dashboard style with `ListView`, `Card`s, and `ListTile`s.
- Improved `settings_page.dart` layout with padding and spacing.
- Fixed lint warnings.
